backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9          # file upload support

# ── Database ───────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9           # PostgreSQL driver

# ── Async task queue ───────────────────────────────────────
celery==5.4.0
redis==5.0.4                     # Redis client + Celery broker
flower==2.0.1                    # Celery monitoring UI

# ── Storage ────────────────────────────────────────────────
boto3==1.34.69                   # AWS S3 / MinIO compatible
minio==7.2.5                     # MinIO SDK (optional alternative)

# ── OCR ────────────────────────────────────────────────────
torch==2.3.0
transformers==4.41.0             # GLM-OCR model loading
Pillow==10.3.0                   # Image preprocessing
pdf2image==1.17.0                # PDF → image conversion

# ── Validation & Config ────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ── Utilities ──────────────────────────────────────────────
python-jose[cryptography]==3.3.0 # JWT (auth, optional)
httpx==0.27.0                    # Async HTTP client (tests)
aiofiles==23.2.1                 # Async file I/O

# ── Dev & Testing ──────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0